Moved PageInfo types to shared GraphQLTypes
