+++ Sources/App/Controllers/GraphQLController.swift
@@ -1,6 +1,5 @@
 import Vapor
 import GraphQL
-import Pagination
 import AnyCodable
 
 class GraphQLController: PigeonController {
@@ -23,11 +22,9 @@ private extension GraphQLController {
         return request.allContentTypes().map { contentTypes in
             var rootFields = [String: GraphQLField]()
 
-            let pageInfo = try self.graphQLPageInfoType()
-
             for type in contentTypes {
                 rootFields[type.plural.camelCase()] = try GraphQLField(
-                    type: type.graphQLType(pageInfo),
+                    type: type.graphQLType(GraphQLPageInfoType),
                     args: type.graphQLPaginationArgs(),
                     resolve: type.rootResolver()
                 )
@@ -40,38 +37,6 @@ private extension GraphQLController {
                 types: SupportedType.graphQLNamedTypes
             )
             return schema
-        }
-    }
-
-    func graphQLPageInfoType() throws -> GraphQLOutputType {
-        var fields = [String: GraphQLField]()
-        fields["current"] = GraphQLField(type: GraphQLInt, resolve: paginationResolver())
-        fields["size"] = GraphQLField(type: GraphQLInt, resolve: paginationResolver())
-        fields["total"] = GraphQLField(type: GraphQLInt, resolve: paginationResolver())
-
-        let pageInfo = try GraphQLObjectType(name: "PageInfo",
-                                             fields: fields)
-        return pageInfo
-    }
-
-    func paginationResolver() -> GraphQLFieldResolve {
-        return { (source, args, context, eventLoopGroup, info) -> EventLoopFuture<Any?> in
-            guard let page = source as? Page<ContentItem> else {
-                throw Abort(.serviceUnavailable)
-            }
-            guard info.path.count > 2 else {
-                throw Abort(.serviceUnavailable)
-            }
-            switch info.path[2].keyValue {
-            case "current":
-                return eventLoopGroup.next().newSucceededFuture(result: page.number)
-            case "total":
-                return eventLoopGroup.next().newSucceededFuture(result: Int(ceil(Float(page.total) / Float(page.size))))
-            case "size":
-                return eventLoopGroup.next().newSucceededFuture(result: page.size)
-            default:
-                throw Abort(.serviceUnavailable)
-            }
         }
     }
